workers/sensory-shield/requirements.txt
websockets==12.0
pyaudiowpatch==0.2.12.3
redis==5.0.3
python-dotenv==1.0.1
